# BackendFinal/TaxCalculationProjectFinal/TaxCalculationProjectFinal/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TaxCalculationProjectFinal, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/TableInfoes": {
      "get": {
        "tags": [
          "TableInfoes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TableInfo"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TableInfoes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TableInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TableInfo"
            }
          }
        }
      }
    },
    "/api/TableInfoes/{id}": {
      "get": {
        "tags": [
          "TableInfoes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TableInfo"
            }
          }
        }
      },
      "put": {
        "tags": [
          "TableInfoes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TableInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "TableInfoes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/TableInfoes/userdetails/{id}": {
      "get": {
        "tags": [
          "TableInfoes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TableInfo"
              }
            }
          }
        }
      }
    },
    "/api/TableInfoes/AllUsersForCA/{id}": {
      "get": {
        "tags": [
          "TableInfoes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TableInfo"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "CharteredAccountant": {
      "type": "object",
      "properties": {
        "caId": {
          "format": "int32",
          "type": "integer"
        },
        "caName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "cpassword": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "caAddress": {
          "type": "string"
        },
        "tableInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TableInfo"
          }
        },
        "userData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserDatum"
          }
        }
      },
      "additionalProperties": false
    },
...
    "TableInfo": {
      "type": "object",
      "properties": {
        "taxId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "caId": {
          "format": "int32",
          "type": "integer"
        },
        "assessmentYear": {
          "type": "string"
        },
        "age": {
          "type": "string"
        },
        "totalAnnualIncome": {
          "format": "double",
          "type": "number"
        },
        "totalDeductions": {
          "format": "double",
          "type": "number"
        },
        "oldRegimeTax": {
          "format": "double",
          "type": "number"
        },
        "newRegimeTax": {
          "format": "double",
          "type": "number"
        },
        "ca": {
          "$ref": "#/definitions/CharteredAccountant"
        },
        "user": {
          "$ref": "#/definitions/UserDatum"
        }
      },
      "additionalProperties": false
    },
    "UserDatum": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "caId": {
          "format": "int32",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "userAddress": {
          "type": "string"
        },
        "upassword": {
          "type": "string"
        },
        "ca": {
          "$ref": "#/definitions/CharteredAccountant"
        },
        "tableInfos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TableInfo"
          }
        }
      },
      "additionalProperties": false
    },
...
  }
}